backend/pyproject.toml
[project]
name = "dsat-backend"
version = "0.1.0"
description = "DSAT question corpus ingestion, annotation, generation, and practice API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.18",
    "anthropic>=0.40",
    "openai>=1.50",
    "httpx>=0.28",
    "pymupdf>=1.25",
    "pillow>=11.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aiosqlite>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
